refactor(creation): reorganize complex components into subfolders
Extract 6 modal-containing components into organized subfolders following
the hybrid component organization pattern documented in CLAUDE.md:

- IdentitiesCard (1653 lines) → identities/
- QualitiesCard (1086 lines) → qualities/
- KnowledgeLanguagesCard (912 lines) → knowledge-languages/
- MagicPathCard (922 lines) → magic-path/
- ContactsCard (758 lines) → contacts/
- MetatypeCard (408 lines) → metatype/

Each subfolder contains types.ts, constants.ts, modal components, and
main component with index.ts for re-exports. Updates CLAUDE.md with
component file organization guidelines.
