+++ components/creation/contacts/constants.ts
@@ -1,0 +1,8 @@
+// SR5 CRB p.387: Connection 1-12, Loyalty 1-6
+// Contact cost = Connection + Loyalty (CRB p.98)
+export const MAX_CONNECTION = 12;
+export const MAX_LOYALTY = 6;
+
+// SR5 CRB p.98: A single contact may not have more than 7 Karma spent on them at character creation
+export const MAX_KARMA_PER_CONTACT = 7;
+export const MIN_KARMA_PER_CONTACT = 2;
