+++ components/creation/identities/index.ts
@@ -1,0 +1,1 @@
+export { IdentitiesCard } from "./IdentitiesCard";
