+++ components/creation/knowledge-languages/index.ts
@@ -1,0 +1,1 @@
+export { KnowledgeLanguagesCard } from "./KnowledgeLanguagesCard";
