+++ components/creation/magic-path/index.ts
@@ -1,0 +1,1 @@
+export { MagicPathCard } from "./MagicPathCard";
