+++ components/creation/metatype/constants.ts
@@ -1,0 +1,11 @@
+// SR5 flavor text for metatypes
+export const METATYPE_DESCRIPTIONS: Record<string, string> = {
+  human:
+    "Adaptable and ambitious, humans are the most common metatype in the Sixth World. Their versatility makes them suited to any role.",
+  elf: "Graceful and long-lived, elves are known for their keen senses and natural charisma. Many gravitate toward social or magical roles.",
+  dwarf:
+    "Stout and resilient, dwarves possess natural resistance to toxins and magic. Their determination is legendary.",
+  ork: "Strong and tough, orks mature quickly and often form tight-knit communities. They excel in physical confrontations.",
+  troll:
+    "Massive and powerful, trolls possess natural dermal armor and extended reach. Their size commands respect and fear.",
+};
