+++ components/creation/metatype/index.ts
@@ -1,0 +1,1 @@
+export { MetatypeCard } from "./MetatypeCard";
